rename useModuleFor -> useModuleInfo
