+++ src/hooks/index.js
@@ -1,3 +1,3 @@
 export { default as useUserTenantPermissions } from './useUserTenantPermissions';
-export { default as useModuleFor } from './useModuleFor';
+export { default as useModuleInfo } from './useModuleInfo';
