+++ src/hooks/useModuleInfo.test.js
@@ -4,7 +4,7 @@ import {
   QueryClientProvider,
 } from 'react-query';
 
-import useModuleFor from './useModuleFor';
+import useModuleInfo from './useModuleInfo';
 import useOkapiKy from '../useOkapiKy';
 
 const response = [
@@ -155,7 +155,7 @@ const wrapper = ({ children }) => (
 );
 
 
-describe('useModuleFor', () => {
+describe('useModuleInfo', () => {
   beforeEach(() => {
     useOkapiKy.get = () => ({
       json: () => console.log({ response })
@@ -164,26 +164,26 @@ describe('useModuleFor', () => {
 
   describe('returns the module-name that provides the interface containing a given path', () => {
     it('handles paths with leading /', async () => {
-      const { result } = renderHook(() => useModuleFor('/users'), { wrapper });
+      const { result } = renderHook(() => useModuleInfo('/users'), { wrapper });
       await waitFor(() => result.current.module.name);
       expect(result.current.module.name).toEqual('mod-users');
     });
 
     it('handles paths without leading /', async () => {
-      const { result } = renderHook(() => useModuleFor('inventory-reports/items-in-transit'), { wrapper });
+      const { result } = renderHook(() => useModuleInfo('inventory-reports/items-in-transit'), { wrapper });
       await waitFor(() => result.current.module.name);
       expect(result.current.module.name).toEqual('mod-circulation');
     });
 
     it('ignores query string', async () => {
-      const { result } = renderHook(() => useModuleFor('/users?query=foo==bar'), { wrapper });
+      const { result } = renderHook(() => useModuleInfo('/users?query=foo==bar'), { wrapper });
       await waitFor(() => result.current.module.name);
       expect(result.current.module.name).toEqual('mod-users');
     });
   });
 
   it('returns undefined given an unmatched path', async () => {
-    const { result } = renderHook(() => useModuleFor('/monkey-bagel'), { wrapper });
+    const { result } = renderHook(() => useModuleInfo('/monkey-bagel'), { wrapper });
     await waitFor(() => result.current.module);
     expect(result.current.module).toBeUndefined();
   });
